requirements.txt
pynput>=1.7.6
pystray>=0.19.4
Pillow>=10.0.0
pywin32>=306
psutil>=5.9.0